+++ src/main/java/school/faang/user_service/controller/event/EventParticipationController.java
@@ -1,37 +1,25 @@
 package school.faang.user_service.controller.event;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
-import school.faang.user_service.entity.User;
-import school.faang.user_service.entity.event.Event;
-import school.faang.user_service.repository.UserRepository;
-import school.faang.user_service.repository.event.EventRepository;
 import school.faang.user_service.service.event.EventParticipationService;
 
-@Component
 @RequiredArgsConstructor
 @Controller
 public class EventParticipationController {
     private final EventParticipationService eventParticipationService;
-    private final UserRepository userRepository;
-    private final EventRepository eventRepository;
 
     public void registerParticipant(long eventId, long userId) {
-        User user = userRepository.findById(userId).orElse(null);
-        Event event = eventRepository.findById(eventId).orElse(null);
-
-        validateParams(user, event);
-
-        eventParticipationService.registerParticipant(event.getId(), user.getId());
+        validateParams(eventId, userId);
+        eventParticipationService.registerParticipant(eventId, userId);
     }
 
-    private static void validateParams(User user, Event event) {
-        if (user == null) {
-            throw new NullPointerException("User not found");
+    private void validateParams(long eventId, long userId) {
+        if (eventId < 0) {
+            throw new IllegalArgumentException("Event not found");
         }
-        if (event == null) {
-            throw new NullPointerException("Event not found");
+        if (userId < 0) {
+            throw new IllegalArgumentException("User not found");
         }
     }
 }
